MRADEV-853_MRADEV-854 Fix Service Routing
Post and album routing was caused by incorrectly trying to get environment variables from the docker container. Ingenious.js is called by the browser, and does not have access to those environment variables. Here, they are set back to the values set in the fabric model, but they are going to be interpreted differently by router-mesh when called.